requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
matplotlib==3.8.2
numpy==1.26.2

opencv_python==4.8.1.78
Pillow==10.1.0
protobuf==4.25.1
PyYAML==6.0.1
Requests==2.31.0
scipy==1.11.4
Cython==3.0.6
scikit-image
coremltools==5.2.0
# torch==1.9.1 # --extra-index-url https://download.pytorch.org/wh1/cu111
# torchvision==0.10.1 # --extra-index-url https://download.pytorch.org/wh1/cu111
lion_pytorch==0.1.2

tqdm
setuptools==58.1.0

# logging -------------------------------------
wandb
tensorboard==2.14.1
tensorboardx>=2.1

# plotting ------------------------------------
seaborn==0.13.0
pandas==2.1.4
ipython

# export --------------------------------------
# coremltools>=4.1
# onnx==1.15.0
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
thop  # FLOPS computation
pycocotools>=2.0  # 







